// src/app/core/models/alternativa.ts
export interface Alternativa{
  idAlternativaDisenio: number;
  idMantenimientoVialEvento: number;
  alternativa: string;
  idTipoSuperficieDisenio: number;
  idTipoIntervencionFinalDisenio: number;
  idTipoMetodologiaDisenio: number;
  idTipoMaterialGranular: number;
  espesorDisenio: number;
  cbrInicialPct: number;
  idTipoGeosinteticos: number;
  idTipoSistemaDrenaje: number;
  observaciones: string;
  seleccionada: string
}